change(electron/main): add underscore to allowed path characters
